refactor(playwright): replace manual steps with util from common test and update for cypress too
