+++ packages/common/test/src/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './stacks-attributes';
+export * from './create-next-app';
