+++ packages/cypress/src/utils/test-utils.ts
@@ -1,4 +1,3 @@
-import { addStacksAttributes } from '@ensono-stacks/test';
 import {
     generateFiles,
     getProjects,
@@ -6,9 +5,6 @@ import {
     offsetFromRoot,
     Tree,
 } from '@nrwl/devkit';
-import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
-import { applicationGenerator } from '@nrwl/next';
-import { Schema as NextSchema } from '@nrwl/next/src/generators/application/schema';
 import path from 'path';
 
 import { CypressGeneratorSchema } from '../generators/init/schema';
@@ -17,22 +13,6 @@ export interface NormalizedSchema extends CypressGeneratorSchema {
     projectName: string;
     projectRoot: string;
     cypressProject: string;
-}
-
-export async function createNextApp(
-    applicationName: string,
-    schema?: Partial<NextSchema>,
-) {
-    const appTree = createTreeWithEmptyWorkspace({ layout: 'apps-libs' });
-    await applicationGenerator(appTree, {
-        name: applicationName,
-        style: 'css',
-        e2eTestRunner: 'none',
-        ...schema,
-    });
-
-    addStacksAttributes(appTree, applicationName);
-    return appTree;
 }
 
 export function checkOneOccurence(array: any[], value: any) {
